[STYLE, REFACTORING] - changed jwt, encryption and password to index.js in their respective folders for a better folder structure. Styled files following StandardJS (those sequelize auto-generated files). Applied Abel's suggestions regarding blank spaces and code implementation
