+++ src/index.js
@@ -1,11 +1,9 @@
-
 import express from 'express'
 import { ApolloServer } from 'apollo-server-express'
 import { createServer } from 'http'
-import bodyParser from 'body-parser'
-
 import schema from './schema'
 import { passportMiddleWare } from './passport'
+
 require('dotenv').config()
 
 const port = process.env.PORT || 3001
