+++ src/jwt/index.js
@@ -1,0 +1,7 @@
+import jsonwebtoken from 'jsonwebtoken'
+
+const expiresIn = process.env.JWT_LIFESPAN || '10d'
+
+export function generateJWT (payload, subject) {
+  return jsonwebtoken.sign(payload, process.env.SECRET, { subject, expiresIn })
+}
