+++ src/schema/resolvers/index.js
@@ -1,14 +1,14 @@
-import path from 'path';
-import fs from 'fs';
-import { merge } from 'lodash';
+import path from 'path'
+import fs from 'fs'
+import { merge } from 'lodash'
 
-let resolvers = {};
+let resolvers = {}
 
 fs
   .readdirSync(__dirname)
   .filter(file => file.indexOf('.') !== 0 && file !== 'index.js')
   .forEach((file) => {
-    resolvers = merge(resolvers, require(path.join(__dirname, file)).default);
-  });
+    resolvers = merge(resolvers, require(path.join(__dirname, file)).default)
+  })
 
-module.exports = resolvers;
+module.exports = resolvers
